requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.9
PyJWT==2.9.0
pytest==8.3.4
python-dotenv==1.0.1
Werkzeug==3.0.4
